backend/requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
ipywidgets
pandas
kagglehub
fastapi
uvicorn
python-multipart
pillow
